flagship/scheduling-assistant/requirements.txt
# Core framework dependency
mcp-agent @ file://../../  # Link to the local mcp-agent project root

# Additional dependencies specific to this project
openai
google-api-python-client>=2.160.0
google-auth>=2.38.0
google-auth-oauthlib>=1.2.1
